[ceph] Refactor ceph plugin to split CrushMap
